Relocate ServiceRegistry tests to tests/services/test_registry (#587)
